ATLEDGE-75 - Time : delayMicroseconds
Implement delayMicroseconds function in Assembly and using the ARC's
instructions timing as a reference.
The instruction's timing was computed with a scope and using infinite
loops, meaning the code most likely used the advantage of ARC'c cache.

Function's accuracy is +- 0.15 microseconds in the above stated conditions.
The accuracy can be affected by cache miss annd/or interrupts handling.
It doesn't disable interrupts and it doesn't use any of the SoC power
management features.

Signed-off-by: Bogdan Pricop <bogdan.pricop@emutex.com>