imu: remove delay between bi-dir SPI transfers causing data loss
A wait for TX to complete, followed then by RX, was causing data
loss due to the fact that it was waiting for the busy status flag
to clear after TX, but this doesn't clear until after RX is also
complete.  The result was data loss due to RX FIFO overflow.
Fixed by disabling the wait between TX and RX steps.

This affected functions requesting more than 8 bytes from the
BMI160, such as getMotion6() and getFIFOBytes() methods.

Signed-off-by: Dan O'Donovan <dan@emutex.com>