Flush wait for transmission to be complete
In previous version, flush wait for the transmission to be complete
 by checking if the transmission holding register (or fifo) is empty.
When this happen, the last byte is still in the shift register,
and has still to be transmitted.
This mean we can't know when the transmission is actually complete,
which is crucial to implement RS485 communication, for example.

This commit change this behaviour and wait for the shift register to be empty,
 which means the transmission is really complete.