Fix UART receiver corruption
When the UART receiver is disconnected for a time shorter than a line break
and longer than a start bit, it interpret that as a received packet, and if
data are transmitted immediatly after, they are corrupted.
This commit is a workarround for that bug, that enable the loopback feature
of the UART when UART is disconnected, so the UART module does not notice
it is disconnected.
A previous workarround for this bug (a forced delay after disconnection) is
not needed anymore and is removed.