Enable interrupt mode of UART tx
Previous version of write will never use the interrupt mode
transmission, because it fails to check if holding register is
empty. This commit fix that, and several bugs related to
interrupt mode serial transmission.
When possible, it uses the tx FIFO, in order to reduce
the number of interrupts called and the overhead.